fix: corrected mistakes & minor changes

+++ app/src/main/java/com/devshish/internship/domain/model/Song.kt
@@ -6,6 +6,6 @@ data class Song(
     val uri: Uri?,
     val title: String,
     val artist: String,
-    val duration: Int,
+    val duration: Long,
     val imageUrl: String?
 )
